Working solution for Log on PSDesiredStateConfiguration 1.1
Deviating Log to the internal resource implementation actually works
even on Windows Powershell (PSDesiredStateConfiguration 1.1) as long as
the module is not in a folder in the user's path (apparently SYSTEM has
full control everywhere, but that's not enough).

Review of the verbose output system: the output is now the same across
all Powershell versions, including Windows Powershell. Redirection and
output analysis is used, especially on Windows Powershell, to trap default
verbose output and reuse it appropriately.